requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-writable-nested==0.7.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
